# Performance optimization for integer value serialization.