--- a/Src/Newtonsoft.Json/JsonTextWriter.cs
+++ b/Src/Newtonsoft.Json/JsonTextWriter.cs
@@ -832,10 +832,10 @@ namespace Newtonsoft.Json
 
         private int WriteNumberToBuffer(ulong value, bool negative)
         {
-            if(value < uint.MaxValue)
+            if (value < uint.MaxValue)
             {
                 // avoid the 64 bit division if possible
-                return WriteNumberToBuffer((uint)value, negative);                
+                return WriteNumberToBuffer((uint)value, negative);
             }
 
             EnsureWriteBuffer();