Replaces a modulus operation with a multiply and subtract, which is measurably faster. Added a benchmark to target the change.

Running the benchmark in this configuration:
.NET Framework 4.7 (CLR 4.0.30319.42000), 64bit RyuJIT-v4.7.2558.0

Before:
 SerializeIntegers | 895.3 us | 1.851 us | 1.546 us |

After:
 SerializeIntegers | 726.8 us | 3.092 us | 2.892 us |