@MarkPflug 2017-12-19
I found more performance, by adding method overloads for int32 values. Turns out 64 bit integer division is *especially* expensive.

On a different machine now, so benchmark baseline numbers are different:

Before:
 SerializeIntegers | 637.2 us | 0.6748 us | 0.5635 us |

After:
SerializeIntegers | 318.8 us | 1.158 us | 1.026 us |

That is bizarrely close to 50%.