@JamesNK 2017-12-19
Looks good, just rename the parameter. Not sure why I had `uvalue` in the original overload.