@MarkPflug 2017-12-19
Another minor tweak occurred to me after I logged off last night. The 64 bit path will now call over to the faster 32 bit overload when possible.